requirements.txt
facenet-pytorch==2.5.3
torch==2.3.0
torchvision==0.18.0
numpy==1.26.4
opencv-python==4.9.0.80
Pillow==10.3.0
scikit-learn==1.4.2
pandas==2.2.2
tqdm==4.66.2